requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiomysql==0.2.0
aiosignal==1.3.1
asynctest==0.13.0
attrs==24.2.0
colorama==0.4.6
discord.py==2.4.0
frozenlist==1.4.1
idna==3.8
iniconfig==2.0.0
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==1.0.1
yarl==1.9.4
